moved home page public puzzles to public puzzles page instead, refactored thunks, backend endpoints to go to public_puzzles instead of just to /puzzles, changed url of public puzzles in App.js as well
