+++ frontend/src/components/styles/PublicPuzzles.css
